fix: Move state initialization into __init__
The old solution used a global constant, which was set at compile time.
Hence the date/timestamp in the default directory name didn't reflect
the load time of the module but rather its compilation time, causing the
dump files to be overwritten in every run.

The new solution replaces a function with a constant Ref, which is
initialized at module load time.